demand_prediction/requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==21.2.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
dataclasses==0.8
decorator==5.0.9
dm-tree==0.1.6
Fiona @ file:///D:/pythonProject/Fiona-1.8.18-cp36-cp36m-win_amd64.whl
flatbuffers==1.12
gast==0.4.0
GDAL @ file:///D:/pythonProject/GDAL-3.1.4-cp36-cp36m-win_amd64.whl
geopandas==0.9.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
importlib-metadata==4.5.0
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
munch==2.5.0
numpy @ file:///D:/pythonProject/numpy-1.19.5%2Bmkl-cp36-cp36m-win_amd64.whl
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==0.25.3
Pillow==8.2.0
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygeos==0.10
pyparsing==2.4.7
pyproj==3.0.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Rtree==0.9.7
scikit-learn==0.24.2
scipy==1.5.4
Shapely==1.7.1
six==1.15.0
sklearn==0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-probability==0.12.2
termcolor==1.1.0
threadpoolctl==2.1.0
torch @ http://download.pytorch.org/whl/cu91/torch-0.4.0-cp36-cp36m-win_amd64.whl
typing-extensions==3.7.4.3
urllib3==1.26.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
